// src/shared/api/generated/models/reportingCatalogDatasetDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { ReportingCatalogDimensionDto } from "./reportingCatalogDimensionDto";
import type { ReportingCatalogMetricDto } from "./reportingCatalogMetricDto";
import type { ReportingReadinessDto } from "./reportingReadinessDto";

export interface ReportingCatalogDatasetDto {
  datasetCode: string;
  datasetRevisionId: string;
  dimensions: ReportingCatalogDimensionDto[];
  metrics: ReportingCatalogMetricDto[];
  name: string;
  readiness: ReportingReadinessDto;
  /** @minimum 1 */
  revision: number;
  semanticDigest: string;
}